# OpenAPI Documentation.json
{"openapi":"3.0.0","servers":[{"url":"http://example.com","description":"","variables":{}}],"info":{"version":"f36e7507-8a46-4ce3-a694-693307598b95","title":"GamesTestProject","description":"","termsOfService":"","contact":{},"license":{"name":""}},"paths":{"/api/v1/games":{"get":{"summary":"GET all elements","operationId":"GETallelements","parameters":[],"responses":{"200":{"description":"","headers":{}}}},"post":{"summary":"CREATE one element","operationId":"CREATEoneelement","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"name":"testName","publisher":"testPublisher","price":"0.0f","rating":"0"}},"example":"{\r\n    \"name\": \"testName\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n}"}}}}},"/api/v1/games/1":{"get":{"summary":"GET one element","operationId":"GEToneelement","parameters":[],"responses":{"200":{"description":"","headers":{}}}},"delete":{"summary":"DELETE one element","operationId":"DELETEoneelement","parameters":[],"responses":{"200":{"description":"","headers":{}}}},"put":{"summary":"UPDATE one element","operationId":"UPDATEoneelement","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"id":1,"name":"testName","publisher":"testPublisher","price":"0.0f","rating":"0"}},"example":"{\r\n    \"id\" : 1,\r\n    \"name\": \"testName\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n}"}}}}},"/api/v1/games/import":{"post":{"summary":"CREATE batch of elements","operationId":"CREATEbatchofelements","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":[{"name":"testName","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName2","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName3","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName4","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName5","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName6","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName7","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName8","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName9","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName10","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName10","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName11","publisher":"testPublisher13","price":"0.0f","rating":"0"},{"name":"testName12","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName12","publisher":"testPublisher1","price":"0.0f","rating":"0"},{"name":"testName12","publisher":"testPublisher13","price":"0.0f","rating":"0"},{"name":"testName14","publisher":"testPublisher46","price":"0.0f","rating":"0"},{"name":"testName156","publisher":"testPublisher15","price":"0.0f","rating":"0"},{"name":"testName16","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName18","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName17","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName18","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName19","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName20","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName21","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName22","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName23","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName24","publisher":"testPublisher","price":"0.0f","rating":"0"},{"name":"testName25","publisher":"testPublisher68","price":"1.0f","rating":"1"},{"name":"testName26","publisher":"testPublisher2","price":"2.0f","rating":"5"},{"name":"testName27","publisher":"testPublisher","price":"0.0f","rating":"0"}]},"example":"[{\r\n    \"name\": \"testName\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName2\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName3\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName4\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName5\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName6\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName7\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName8\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName9\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName10\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName10\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName11\",\r\n    \"publisher\":\"testPublisher13\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName12\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName12\",\r\n    \"publisher\":\"testPublisher1\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName12\",\r\n    \"publisher\":\"testPublisher13\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName14\",\r\n    \"publisher\":\"testPublisher46\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName156\",\r\n    \"publisher\":\"testPublisher15\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName16\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName18\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName17\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName18\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName19\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName20\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName21\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName22\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName23\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName24\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n},{\r\n    \"name\": \"testName25\",\r\n    \"publisher\":\"testPublisher68\",\r\n    \"price\":\"1.0f\",\r\n    \"rating\":\"1\"\r\n},{\r\n    \"name\": \"testName26\",\r\n    \"publisher\":\"testPublisher2\",\r\n    \"price\":\"2.0f\",\r\n    \"rating\":\"5\"\r\n},{\r\n    \"name\": \"testName27\",\r\n    \"publisher\":\"testPublisher\",\r\n    \"price\":\"0.0f\",\r\n    \"rating\":\"0\"\r\n}]"}}}}},"/api/v1/games/count":{"get":{"summary":"COUNT elements in database","operationId":"COUNTelementsindatabase","parameters":[],"responses":{"200":{"description":"","headers":{}}}}},"/api/v1/games/batch":{"delete":{"summary":"DELETE all elements","operationId":"DELETEallelements","parameters":[],"responses":{"200":{"description":"","headers":{}}}}}},"components":{},"security":[],"tags":[],"externalDocs":{"url":"","description":""},"warnings":["Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games","Two GET or two POST methods for the same path are not allowed in OpenAPI specification : getlocalhost:8080/api/v1/games"]}